requirements.txt
fastapi==0.116.2
requests==2.32.5
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
groq==0.32.0
pymupdf==1.26.4
openai==1.108.0